Arcpy script written to calculate risk level of NC swine lagoons
This script uses the Arcpy Library from ArcGIS to calculate which census
tracts in North Carolina are at the greatest risk from swine lagoon air
pollution. The risk level for each tract was calculated by first finding
all swine lagoons within 5 miles of the centroid of the tract. For each
lagoon, the estimated number of hogs whose fecal waste the lagoon
contains was divided by the distance of the lagoon from the centroid.
The sum of these quotients was assigned as the risk level of the census
tract.